requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
aiogram==2.9.2
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.42.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
iniconfig==1.0.0
more-itertools==8.2.0
msgpack==0.6.2
multidict==4.7.6
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-django==3.9.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.8
yarl==1.5.1
zipp==3.0.0
zope.interface==5.0.2
